+++ src/detections/rule/matchers.rs
@@ -1,3 +1,4 @@
+use base64::{engine::general_purpose, Engine as _};
 use nested::Nested;
 use regex::Regex;
 use std::cmp::Ordering;
@@ -193,7 +194,7 @@ enum FastMatch {
 /// ワイルドカードの処理やパイプ
 pub struct DefaultMatcher {
     re: Option<Regex>,
-    case_ignore_fast_match: Option<FastMatch>,
+    fast_match: Option<Vec<FastMatch>>,
     pipes: Vec<PipeElement>,
     key_list: Nested<String>,
 }
@@ -202,7 +203,7 @@ impl DefaultMatcher {
     pub fn new() -> DefaultMatcher {
         DefaultMatcher {
             re: Option::None,
-            case_ignore_fast_match: Option::None,
+            fast_match: Option::None,
             pipes: Vec::new(),
             key_list: Nested::<String>::new(),
         }
@@ -266,7 +267,7 @@ impl DefaultMatcher {
     }
 
     // ワイルドカードマッチを高速なstd::stringのlen/starts_with/ends_withに変換するための関数
-    fn convert_to_fast_match(s: &str) -> Option<FastMatch> {
+    fn convert_to_fast_match(s: &str, ignore_case: bool) -> Option<Vec<FastMatch>> {
         let wildcard_count = s.chars().filter(|c| *c == '*').count();
         let is_literal_asterisk = |s: &str| s.ends_with(r"\*") && !s.ends_with(r"\\*");
         if s.contains('?') || s.ends_with(r"\\\*") || (!s.is_ascii() && s.contains('*')) {
@@ -278,23 +279,30 @@ impl DefaultMatcher {
             && !is_literal_asterisk(s)
         {
             // *が先頭と末尾だけは、containsに変換
-            return Some(FastMatch::Contains(
-                s.to_lowercase().replacen('*', "", 2).replace(r"\\", r"\"),
-            ));
+            if ignore_case {
+                return Some(vec![FastMatch::Contains(
+                    s.to_lowercase().replacen('*', "", 2).replace(r"\\", r"\"),
+                )]);
+            }
+            return Some(vec![FastMatch::Contains(
+                s.replacen('*', "", 2).replace(r"\\", r"\"),
+            )]);
         } else if s.starts_with('*') && wildcard_count == 1 {
             // *が先頭は、ends_withに変換
-            return Some(FastMatch::EndsWith(s.replace('*', "").replace(r"\\", r"\")));
+            return Some(vec![FastMatch::EndsWith(
+                s.replace('*', "").replace(r"\\", r"\"),
+            )]);
         } else if s.ends_with('*') && wildcard_count == 1 && !is_literal_asterisk(s) {
             // *が末尾は、starts_withに変換
-            return Some(FastMatch::StartsWith(
+            return Some(vec![FastMatch::StartsWith(
                 s.replace('*', "").replace(r"\\", r"\"),
-            ));
+            )]);
         } else if s.contains('*') {
             // *が先頭・末尾以外にあるパターンは、starts_with/ends_withに変換できないため、正規表現マッチのみ
             return None;
         }
         // *を含まない場合は、文字列長マッチに変換
-        Some(FastMatch::Exact(s.replace(r"\\", r"\")))
+        Some(vec![FastMatch::Exact(s.replace(r"\\", r"\"))])
     }
 }
 
@@ -354,12 +362,14 @@ impl LeafMatcher for DefaultMatcher {
         if n == 0 {
             // パイプがないケース
             if let Some(val) = select_value.as_str() {
-                self.case_ignore_fast_match = Self::convert_to_fast_match(val);
+                self.fast_match = Self::convert_to_fast_match(val, true);
             };
-            if matches!(
-                &self.case_ignore_fast_match,
-                Some(FastMatch::Exact(_)) | Some(FastMatch::Contains(_))
-            ) && !self.key_list.is_empty()
+            if self.fast_match.is_some()
+                && matches!(
+                    &self.fast_match.as_ref().unwrap()[0],
+                    FastMatch::Exact(_) | FastMatch::Contains(_)
+                )
+                && !self.key_list.is_empty()
             {
                 // FastMatch::Exact検索に置き換えられたときは正規表現は不要
                 return Result::Ok(());
@@ -367,21 +377,83 @@ impl LeafMatcher for DefaultMatcher {
         } else if n == 1 {
             // パイプがあるケース
             if let Some(val) = select_value.as_str() {
-                self.case_ignore_fast_match = match &self.pipes[0] {
+                self.fast_match = match &self.pipes[0] {
                     PipeElement::Startswith => {
-                        Self::convert_to_fast_match(format!("{}{}", val, '*').as_str())
+                        Self::convert_to_fast_match(format!("{}{}", val, '*').as_str(), true)
                     }
                     PipeElement::Endswith => {
-                        Self::convert_to_fast_match(format!("{}{}", '*', val).as_str())
+                        Self::convert_to_fast_match(format!("{}{}", '*', val).as_str(), true)
                     }
                     PipeElement::Contains => {
-                        Self::convert_to_fast_match(format!("{}{}{}", '*', val, '*').as_str())
+                        Self::convert_to_fast_match(format!("{}{}{}", '*', val, '*').as_str(), true)
                     }
                     _ => None,
                 };
             }
-        } else if n >= 2 {
-            // 現状では複数のパイプは対応していない
+        } else if n == 2
+            && (self.pipes[0] == PipeElement::Base64offset
+                && self.pipes[1] == PipeElement::Contains)
+        {
+            // 現状では複数のパイプへの対応は|base64offset|containsの場合のみ
+            if let Some(val) = select_value.as_str() {
+                let val_byte = val.as_bytes();
+                let mut fastmatches = vec![];
+                for i in 0..3 {
+                    let mut b64_result = vec![];
+                    let mut target_byte = vec![];
+                    target_byte.resize_with(i, || 0b0);
+                    target_byte.extend_from_slice(val_byte);
+                    b64_result.resize_with(target_byte.len() * 4 / 3 + 4, || 0b0);
+                    general_purpose::STANDARD
+                        .encode_slice(target_byte, &mut b64_result)
+                        .ok();
+                    let convstr_b64 = String::from_utf8(b64_result);
+                    if let Ok(b64_str) = convstr_b64 {
+                        // ここでContainsのfastmatch対応を行う
+                        let filtered_null_chr = b64_str.replace('\0', "");
+                        let b64_offset_contents = match b64_str.find('=').unwrap_or_default() % 4 {
+                            2 => {
+                                if i == 0 {
+                                    filtered_null_chr[..filtered_null_chr.len() - 3].to_string()
+                                } else {
+                                    filtered_null_chr[(i + 1)..filtered_null_chr.len() - 3]
+                                        .to_string()
+                                }
+                            }
+                            3 => {
+                                if i == 0 {
+                                    filtered_null_chr[..filtered_null_chr.len() - 2].to_string()
+                                } else {
+                                    filtered_null_chr.replace('\0', "")
+                                        [(i + 1)..filtered_null_chr.len() - 2]
+                                        .to_string()
+                                }
+                            }
+                            _ => {
+                                if i == 0 {
+                                    filtered_null_chr
+                                } else {
+                                    filtered_null_chr[(i + 1)..].to_string()
+                                }
+                            }
+                        };
+                        if let Some(fm) =
+                            Self::convert_to_fast_match(&format!("*{b64_offset_contents}*"), false)
+                        {
+                            fastmatches.extend(fm);
+                        }
+                    } else {
+                        err_msges.push(format!(
+                            "Failed base64 encoding: {}",
+                            convstr_b64.unwrap_err()
+                        ));
+                    }
+                }
+                if !fastmatches.is_empty() {
+                    self.fast_match = Some(fastmatches);
+                }
+            }
+        } else {
             let errmsg = format!(
                 "Multiple pipe elements cannot be used. key:{}",
                 utils::concat_selection_key(key_list)
@@ -430,12 +502,12 @@ impl LeafMatcher for DefaultMatcher {
 
         // yamlにnullが設定されていた場合
         // keylistが空(==JSONのgrep検索)の場合、無視する。
-        if self.key_list.is_empty() && self.re.is_none() && self.case_ignore_fast_match.is_none() {
+        if self.key_list.is_empty() && self.re.is_none() && self.fast_match.is_none() {
             return false;
         }
 
         // yamlにnullが設定されていた場合
-        if self.re.is_none() && self.case_ignore_fast_match.is_none() {
+        if self.re.is_none() && self.fast_match.is_none() {
             // レコード内に対象のフィールドが存在しなければ検知したものとして扱う
             for v in self.key_list.iter() {
                 if recinfo.get_value(v).is_none() {
@@ -453,12 +525,19 @@ impl LeafMatcher for DefaultMatcher {
         if self.key_list.is_empty() {
             // この場合ただのgrep検索なので、ただ正規表現に一致するかどうか調べればよいだけ
             return self.re.as_ref().unwrap().is_match(event_value_str);
-        } else if let Some(fast_matcher) = &self.case_ignore_fast_match {
-            let fast_match_result = match fast_matcher {
-                FastMatch::Exact(s) => Some(Self::eq_ignore_case(event_value_str, s)),
-                FastMatch::StartsWith(s) => Self::starts_with_ignore_case(event_value_str, s),
-                FastMatch::EndsWith(s) => Self::ends_with_ignore_case(event_value_str, s),
-                FastMatch::Contains(s) => Some(event_value_str.to_lowercase().contains(s)),
+        } else if let Some(fast_matcher) = &self.fast_match {
+            let fast_match_result = if fast_matcher.len() == 1 {
+                match &fast_matcher[0] {
+                    FastMatch::Exact(s) => Some(Self::eq_ignore_case(event_value_str, s)),
+                    FastMatch::StartsWith(s) => Self::starts_with_ignore_case(event_value_str, s),
+                    FastMatch::EndsWith(s) => Self::ends_with_ignore_case(event_value_str, s),
+                    FastMatch::Contains(s) => Some(event_value_str.to_lowercase().contains(s)),
+                }
+            } else {
+                Some(fast_matcher.iter().any(|fm| match fm {
+                    FastMatch::Contains(s) => event_value_str.contains(s),
+                    _ => false,
+                }))
             };
             if let Some(is_match) = fast_match_result {
                 return is_match;
@@ -471,6 +550,7 @@ impl LeafMatcher for DefaultMatcher {
 
 /// パイプ(|)で指定される要素を表すクラス。
 /// 要リファクタリング
+#[derive(PartialEq)]
 enum PipeElement {
     Startswith,
     Endswith,
@@ -479,6 +559,7 @@ enum PipeElement {
     Wildcard,
     EqualsField(String),
     Endswithfield(String),
+    Base64offset,
 }
 
 impl PipeElement {
@@ -490,6 +571,7 @@ impl PipeElement {
             "re" => Option::Some(PipeElement::Re),
             "equalsfield" => Option::Some(PipeElement::EqualsField(pattern.to_string())),
             "endswithfield" => Option::Some(PipeElement::Endswithfield(pattern.to_string())),
+            "base64offset" => Option::Some(PipeElement::Base64offset),
             _ => Option::None,
         };
 
@@ -800,11 +882,13 @@ mod tests {
             assert!(matcher.is::<DefaultMatcher>());
             let matcher = matcher.downcast_ref::<DefaultMatcher>().unwrap();
 
-            assert!(matcher.case_ignore_fast_match.is_some());
-            let fast_match = matcher.case_ignore_fast_match.as_ref().unwrap();
+            assert!(matcher.fast_match.is_some());
+            let fast_match = matcher.fast_match.as_ref().unwrap();
             assert_eq!(
                 *fast_match,
-                FastMatch::Exact("Microsoft-Windows-PowerShell/Operational".to_string())
+                vec![FastMatch::Exact(
+                    "Microsoft-Windows-PowerShell/Operational".to_string()
+                )]
             );
         }
 
@@ -823,7 +907,7 @@ mod tests {
             let matcher = child_node.matcher.as_ref().unwrap();
             assert!(matcher.is::<DefaultMatcher>());
             let matcher = matcher.downcast_ref::<DefaultMatcher>().unwrap();
-            assert!(matcher.case_ignore_fast_match.is_none());
+            assert!(matcher.fast_match.is_none());
         }
 
         // ContextInfo
@@ -846,11 +930,11 @@ mod tests {
             let hostapp_en_matcher = hostapp_en_matcher.as_ref().unwrap();
             assert!(hostapp_en_matcher.is::<DefaultMatcher>());
             let hostapp_en_matcher = hostapp_en_matcher.downcast_ref::<DefaultMatcher>().unwrap();
-            assert!(hostapp_en_matcher.case_ignore_fast_match.is_some());
-            let fast_match = hostapp_en_matcher.case_ignore_fast_match.as_ref().unwrap();
+            assert!(hostapp_en_matcher.fast_match.is_some());
+            let fast_match = hostapp_en_matcher.fast_match.as_ref().unwrap();
             assert_eq!(
                 *fast_match,
-                FastMatch::Exact("Host Application".to_string())
+                vec![FastMatch::Exact("Host Application".to_string())]
             );
 
             // LeafSelectionNodeである、ホスト アプリケーションノードが正しいことを確認
@@ -863,11 +947,11 @@ mod tests {
             let hostapp_jp_matcher = hostapp_jp_matcher.as_ref().unwrap();
             assert!(hostapp_jp_matcher.is::<DefaultMatcher>());
             let hostapp_jp_matcher = hostapp_jp_matcher.downcast_ref::<DefaultMatcher>().unwrap();
-            assert!(hostapp_jp_matcher.case_ignore_fast_match.is_some());
-            let fast_match = hostapp_jp_matcher.case_ignore_fast_match.as_ref().unwrap();
+            assert!(hostapp_jp_matcher.fast_match.is_some());
+            let fast_match = hostapp_jp_matcher.fast_match.as_ref().unwrap();
             assert_eq!(
                 *fast_match,
-                FastMatch::Exact("ホスト アプリケーション".to_string())
+                vec![FastMatch::Exact("ホスト アプリケーション".to_string())]
             );
         }
 
@@ -2223,39 +2307,79 @@ mod tests {
     }
     #[test]
     fn test_convert_to_fast_match() {
-        assert_eq!(DefaultMatcher::convert_to_fast_match("ab?"), None);
-        assert_eq!(DefaultMatcher::convert_to_fast_match("a*c"), None);
-        assert_eq!(DefaultMatcher::convert_to_fast_match("*a*b"), None);
-        assert_eq!(DefaultMatcher::convert_to_fast_match("*a*b*"), None);
-        assert_eq!(DefaultMatcher::convert_to_fast_match(r"a\*"), None);
-        assert_eq!(DefaultMatcher::convert_to_fast_match(r"a\\\*"), None);
+        assert_eq!(DefaultMatcher::convert_to_fast_match("ab?", true), None);
+        assert_eq!(DefaultMatcher::convert_to_fast_match("a*c", true), None);
+        assert_eq!(DefaultMatcher::convert_to_fast_match("*a*b", true), None);
+        assert_eq!(DefaultMatcher::convert_to_fast_match("*a*b*", true), None);
+        assert_eq!(DefaultMatcher::convert_to_fast_match(r"a\*", true), None);
+        assert_eq!(DefaultMatcher::convert_to_fast_match(r"a\\\*", true), None);
         assert_eq!(
-            DefaultMatcher::convert_to_fast_match("abc*").unwrap(),
-            FastMatch::StartsWith("abc".to_string())
+            DefaultMatcher::convert_to_fast_match("abc*", true).unwrap(),
+            vec![FastMatch::StartsWith("abc".to_string())]
         );
         assert_eq!(
-            DefaultMatcher::convert_to_fast_match(r"abc\\*").unwrap(),
-            FastMatch::StartsWith(r"abc\".to_string())
+            DefaultMatcher::convert_to_fast_match(r"abc\\*", true).unwrap(),
+            vec![FastMatch::StartsWith(r"abc\".to_string())]
         );
         assert_eq!(
-            DefaultMatcher::convert_to_fast_match("*abc").unwrap(),
-            FastMatch::EndsWith("abc".to_string())
+            DefaultMatcher::convert_to_fast_match("*abc", true).unwrap(),
+            vec![FastMatch::EndsWith("abc".to_string())]
         );
         assert_eq!(
-            DefaultMatcher::convert_to_fast_match("*abc*").unwrap(),
-            FastMatch::Contains("abc".to_string())
+            DefaultMatcher::convert_to_fast_match("*abc*", true).unwrap(),
+            vec![FastMatch::Contains("abc".to_string())]
         );
         assert_eq!(
-            DefaultMatcher::convert_to_fast_match("abc").unwrap(),
-            FastMatch::Exact("abc".to_string())
+            DefaultMatcher::convert_to_fast_match("abc", true).unwrap(),
+            vec![FastMatch::Exact("abc".to_string())]
         );
         assert_eq!(
-            DefaultMatcher::convert_to_fast_match("あいう").unwrap(),
-            FastMatch::Exact("あいう".to_string())
+            DefaultMatcher::convert_to_fast_match("あいう", true).unwrap(),
+            vec![FastMatch::Exact("あいう".to_string())]
         );
         assert_eq!(
-            DefaultMatcher::convert_to_fast_match(r"\\\\127.0.0.1\\").unwrap(),
-            FastMatch::Exact(r"\\127.0.0.1\".to_string())
+            DefaultMatcher::convert_to_fast_match(r"\\\\127.0.0.1\\", true).unwrap(),
+            vec![FastMatch::Exact(r"\\127.0.0.1\".to_string())]
         );
+    }
+
+    #[test]
+    fn test_base64offset_contains() {
+        // base64offset|containsのマッチ
+        let rule_str = r#"
+        enabled: true
+        detection:
+            selection:
+                Payload|base64offset|contains:
+                    - "http://"
+        details: 'command=%CommandLine%'
+        "#;
+
+        let record_json_str = r#"{
+            "Event": {"System": {"EventID": 4103, "Channel": "Security", "Computer": "Tester"}, "EventData":{"Payload": "aHR0cDovL"}},
+            "Event_attributes": {"xmlns": "http://schemas.microsoft.com/win/2004/08/events/event"}
+        }"#;
+
+        check_select(rule_str, record_json_str, true);
+    }
+
+    #[test]
+    fn test_base64offset_contains_not_match() {
+        // base64offset|containsのマッチしないパターン
+        let rule_str = r#"
+        enabled: true
+        detection:
+            selection:
+                Payload|base64offset|contains:
+                    - "test"
+        details: 'command=%CommandLine%'
+        "#;
+
+        let record_json_str = r#"{
+            "Event": {"System": {"EventID": 4103, "Channel": "Security", "Computer": "Tester"}, "EventData":{"Payload": "aHR0cDovL"}},
+            "Event_attributes": {"xmlns": "http://schemas.microsoft.com/win/2004/08/events/event"}
+        }"#;
+
+        check_select(rule_str, record_json_str, false);
     }
 }